Refactored spec tests
The responsiblity of the Game class changed, it no longer does the
marking, that is handled by the Marker class.  The game class just
outputs the mark.  So we make a new marker_spec file that tests the
match counts for exact and number matches, and removed the testing
of individual marks from the game_spec file.  In their place we just
put a test that verifies the mark is output by the game after a
guess is made.